site reorganisation - splitting larger components and reformatting
